Add files via upload
I have made a few changes trying to get the subCategory to work. If you click on any of the categories you will be taken to the subCategoriesScreen. Each sub-category needs to have its own circle icons along the top. I made that work by creating a JSON for each of these that follows the route you added named "categoryInfoData" that begins at line 373 of data/business-data.
     The only question I have...If you click Dining, then click Dining 1, it will show data from a JSON I created in business-data named CATEGORIESINFODINING (line 50). But, I would like to have the title to be shown as the header title. (if you clicked on "Dining 1" the title will read "Dining 1", below the color banner). I cannot get that information to pass through to my useLayoutEffect code that begins on line 9 of "InfoDetailScreenCategories.js"
    Note: In the JSON you will notice two names. The first name is used beside the circle icon and often will have a return in it (see line 41 of business-data). The second line is used for the header title. In the models file (homeInfoBubbles.js) this title is listed as infoTitle2.
   Thanks for any solution you can find.